requirements.txt
datasets==2.6.1
numpy==2.0.1
pandas==2.2.2
PyYAML==6.0.1
torch==1.13.1
torchtext==0.14.1
tqdm==4.65.0
transformers==4.24.0
